requirements.txt
pdfplumber>=0.10.0
numpy>=1.26.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
sentence-transformers>=2.7.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.1.3
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
